Add tox -elinters support
This is the first entry point to use tox.ini, this allows us drop the
dependency on sf-jobs, since it does most of this.  For now, we are
doing flake8, however we really need to have a discussion on the PTI for
ansible-network. To help ensure all roles are testing the same way.

Depends-On: https://github.com/ansible-network/ansible-zuul-jobs/pull/39
Signed-off-by: Paul Belanger <pabelanger@redhat.com>